lestarch: UT and compilation fixes for types refactor

+++ Fw/Types/DefaultTypes.hpp
@@ -55,9 +55,9 @@ const PlatformSizeType PlatformSizeType_MAX = PlatformUIntType_MAX;
 * Default implementation for pointers stored as integers
 */
 #ifndef PLATFORM_POINTER_CAST_TYPE_DEFINED
-typedef uintptr_t PlatformPointerCastType;
-const PlatformPointerCastType PlatformPointerCastType_MIN = std::numeric_limits<uintptr_t>::min();
-const PlatformPointerCastType PlatformPointerCastType_MAX = std::numeric_limits<uintptr_t>::max();
+typedef uint64_t PlatformPointerCastType;
+const PlatformPointerCastType PlatformPointerCastType_MIN = std::numeric_limits<uint64_t>::min();
+const PlatformPointerCastType PlatformPointerCastType_MAX = std::numeric_limits<uint64_t>::max();
 #define PLATFORM_POINTER_CAST_TYPE_DEFINED
 #define PRI_PlatformPointerCastType PRIx64
 #endif
